backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Voice Processing
torch>=2.1.0
torchaudio>=2.1.0
transformers>=4.35.0
librosa>=0.10.1
openai-whisper>=20231117
datasets>=2.14.0
sentencepiece>=0.1.99

# Text Processing
sentence-transformers>=2.2.2

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1
pathlib

# Audio Processing
soundfile>=0.12.1
scipy>=1.11.0